move tracer tests to dedicated folder
