+++ src/tests/mod.rs
@@ -1,0 +1,2 @@
+mod goal_based_tests;
+mod tracer_tests;
